re-structured

+++ controllers/groups.controller.js
@@ -1,4 +1,4 @@
-const db = require("../../models");
+const db = require("../models");
 
 class GroupController {
   constructor() {}
@@ -48,26 +48,76 @@ class GroupController {
     }
   }
   static async createGroup(req, res) {
-    const { name, features } = req.body;
+    const { name, roles, users, features } = req.body;
     try {
-      const group = await db.Role.create({ name });
-      const featureCollection = features.map(async (feature) => {
-        await db.Feature.findOne({
-          where: {
-            [db.Sequelize.and]: [
-              { entityType: feature.entityType },
-              { name: feature.name },
-            ],
-          },
-        });
-      });
-      featureCollection.map(async (feature) => {
-        await db.FeaturePerms.create({
-          entityId: role.dataValues.id,
-          featureId: feature.id,
-          entityName: "Groups",
-        });
-      });
+      const group = await db.Group.create({ name });
+
+      if (roles && roles.length > 0) {
+        const rolesCollection = await Promise.all(
+          roles.map(async (role) => {
+            return await db.Role.findOne({
+              where: {
+                id: role.id,
+              },
+            });
+          })
+        );
+
+        await Promise.all(
+          rolesCollection.map(async (role) => {
+            await db.RoleGroup.create({
+              roleId: role.id,
+              groupId: group.dataValues.id,
+            });
+          })
+        );
+      }
+
+      if (users && users.length > 0) {
+        const usersCollection = await Promise.all(
+          users.map(async (user) => {
+            return await db.User.findOne({
+              where: {
+                id: user.id,
+              },
+            });
+          })
+        );
+        await Promise.all(
+          usersCollection.map(async (user) => {
+            await db.UserGroup.create({
+              userId: user.id,
+              groupId: group.dataValues.id,
+            });
+          })
+        );
+      }
+
+      if (features && features.length > 0) {
+        const featureCollection = await Promise.all(
+          features.map(async (feature) => {
+            return await db.Feature.findOne({
+              where: {
+                [db.Sequelize.and]: [
+                  { entityType: feature.entityType },
+                  { name: feature.name },
+                ],
+              },
+            });
+          })
+        );
+
+        await Promise.all(
+          featureCollection.map(async (feature) => {
+            await db.FeaturePerms.create({
+              entityId: group.dataValues.id,
+              featureId: feature.id,
+              entityName: "Groups",
+            });
+          })
+        );
+      }
+
       res.json({
         status: "success",
       });
@@ -78,6 +128,7 @@ class GroupController {
       });
     }
   }
+
   static async updateGroup(req, res) {
     const id = req.params.groupId;
     const { name, features } = req.body;
@@ -127,15 +178,23 @@ class GroupController {
     }
   }
   static async deleteGroup(req, res) {
-    const id = req.params.groupdId;
+    const id = req.params.groupId;
     try {
-      await db.FeaturePerms.destroy({
+      const featureperm = await db.FeaturePerms.findOne({
         where: {
           [db.Sequelize.Op.and]: [{ entityId: id }, { entityName: "Groups" }],
         },
       });
-      await db.UserGroup.destroy({ where: { groupId: id } });
-      await db.RoleGroup.destroy({ where: { groupId: id } });
+      if (featureperm)
+        await db.FeaturePerms.destroy({
+          where: {
+            [db.Sequelize.Op.and]: [{ entityId: id }, { entityName: "Groups" }],
+          },
+        });
+      const usergroup = await db.UserGroup.findOne({ where: { groupId: id } });
+      if (usergroup) await db.UserGroup.destroy({ where: { groupId: id } });
+      const rolegroup = await db.RoleGroup.findOne({ where: { groupId: id } });
+      if (rolegroup) await db.RoleGroup.destroy({ where: { groupId: id } });
       await db.Group.destroy({
         where: { id: id },
       });
